+++ operator/buffer/memory.go
@@ -68,6 +68,7 @@ type MemoryBuffer struct {
 	sem           *semaphore.Weighted
 	maxChunkDelay time.Duration
 	maxChunkSize  uint
+	reconfigMutex sync.RWMutex
 }
 
 // Add inserts an entry into the memory database, blocking until there is space
@@ -105,7 +106,7 @@ func (m *MemoryBuffer) Read(dst []*entry.Entry) (Clearer, int, error) {
 
 // ReadChunk is a thin wrapper around ReadWait that simplifies the call at the expense of an extra allocation
 func (m *MemoryBuffer) ReadChunk(ctx context.Context) ([]*entry.Entry, Clearer, error) {
-	entries := make([]*entry.Entry, m.maxChunkSize)
+	entries := make([]*entry.Entry, m.MaxChunkSize())
 	for {
 		select {
 		case <-ctx.Done():
@@ -113,7 +114,7 @@ func (m *MemoryBuffer) ReadChunk(ctx context.Context) ([]*entry.Entry, Clearer, 
 		default:
 		}
 
-		ctx, cancel := context.WithTimeout(ctx, m.maxChunkDelay)
+		ctx, cancel := context.WithTimeout(ctx, m.MaxChunkDelay())
 		defer cancel()
 		flushFunc, n, err := m.ReadWait(ctx, entries)
 		if n > 0 {
@@ -143,6 +144,30 @@ func (m *MemoryBuffer) ReadWait(ctx context.Context, dst []*entry.Entry) (Cleare
 	}
 
 	return m.newClearer(inFlightIDs[:i]), i, nil
+}
+
+func (m *MemoryBuffer) MaxChunkSize() uint {
+	m.reconfigMutex.RLock()
+	defer m.reconfigMutex.RUnlock()
+	return m.maxChunkSize
+}
+
+func (m *MemoryBuffer) MaxChunkDelay() time.Duration {
+	m.reconfigMutex.RLock()
+	defer m.reconfigMutex.RUnlock()
+	return m.maxChunkDelay
+}
+
+func (m *MemoryBuffer) SetMaxChunkSize(size uint) {
+	m.reconfigMutex.Lock()
+	m.maxChunkSize = size
+	m.reconfigMutex.Unlock()
+}
+
+func (m *MemoryBuffer) SetMaxChunkDelay(delay time.Duration) {
+	m.reconfigMutex.Lock()
+	m.maxChunkDelay = delay
+	m.reconfigMutex.Unlock()
 }
 
 type memoryClearer struct {