+++ operator/builtin/input/tcp/tcp.go
@@ -7,6 +7,7 @@ import (
 	"crypto/tls"
 	"fmt"
 	"net"
+	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -45,6 +46,7 @@ type TCPInputConfig struct {
 	MaxBufferSize helper.ByteSize `json:"max_buffer_size,omitempty" yaml:"max_buffer_size,omitempty"`
 	ListenAddress string          `json:"listen_address,omitempty" yaml:"listen_address,omitempty"`
 	TLS           TLSConfig       `json:"tls,omitempty" yaml:"tls,omitempty"`
+	AddLabels     bool            `json:"add_labels,omitempty" yaml:"add_labels,omitempty"`
 }
 
 // TLSConfig is the configuration for a TLS listener
@@ -106,6 +108,7 @@ func (c TCPInputConfig) Build(context operator.BuildContext) ([]operator.Operato
 		InputOperator: inputOperator,
 		address:       c.ListenAddress,
 		maxBufferSize: int(c.MaxBufferSize),
+		addLabels:     c.AddLabels,
 		tlsEnable:     c.TLS.Enable,
 		tlsKeyPair:    cert,
 		backoff: backoff.Backoff{
@@ -123,6 +126,7 @@ type TCPInput struct {
 	helper.InputOperator
 	address       string
 	maxBufferSize int
+	addLabels     bool
 	tlsEnable     bool
 	tlsKeyPair    tls.Certificate
 	backoff       backoff.Backoff
@@ -228,6 +232,20 @@ func (t *TCPInput) goHandleMessages(ctx context.Context, conn net.Conn, cancel c
 				t.Errorw("Failed to create entry", zap.Error(err))
 				continue
 			}
+
+			if t.addLabels {
+				entry.AddLabel("net.transport", "IP.TCP")
+				if addr, ok := conn.RemoteAddr().(*net.TCPAddr); ok {
+					entry.AddLabel("net.peer.ip", addr.IP.String())
+					entry.AddLabel("net.peer.port", strconv.FormatInt(int64(addr.Port), 10))
+				}
+
+				if addr, ok := conn.LocalAddr().(*net.TCPAddr); ok {
+					entry.AddLabel("net.host.ip", addr.IP.String())
+					entry.AddLabel("net.host.port", strconv.FormatInt(int64(addr.Port), 10))
+				}
+			}
+
 			t.Write(ctx, entry)
 		}
 		if err := scanner.Err(); err != nil {