+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.14.2] - 2021-05-24
+
+### Changed
+- Make buffer max chunk delay reconfigurable on the fly [PR313](https://github.com/observIQ/stanza/pull/313)
+
+## [0.14.1] - 2021-05-20
+
+### Added
+- Added optional network metadata labels to tcp / udp operators [PR302](https://github.com/observIQ/stanza/pull/302)
+- Added AWS Cloudwatch Logs input operator [PR289](https://github.com/observIQ/stanza/pull/289)
+
 ## [0.14.0] - 2021-05-07
 
 ### Added